requirements.txt
# Core packages
numpy>=1.20.0
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.6.0
Pillow>=9.2.0
matplotlib>=3.5.3
pandas>=1.4.3
scikit-learn>=1.1.2
tensorboard>=2.10.0

# YOLO implementations
ultralytics>=8.0.0  # For YOLOv8
# For other YOLO versions we'll use specific repositories or implementations

# Data processing and augmentation
albumentations>=1.3.0
pycocotools>=2.0.6
roboflow>=0.2.29

# Utilities
tqdm>=4.64.0
PyYAML>=6.0
seaborn>=0.12.0
ipykernel>=6.15.1

# Weights & Biases for experiment tracking
wandb>=0.13.4 